Reorganized ContentBlock Components
