+++ app/routes/page-builder/components/content-block/feature-section.tsx
@@ -1,0 +1,81 @@
+import { FC } from "react";
+import { cn, parseRockKeyValueList } from "~/lib/utils";
+import { Button } from "~/primitives/button/button.primitive";
+import HTMLRenderer from "~/primitives/html-renderer";
+import { ContentBlockData } from "../../types";
+
+export const FeatureImage: FC<{ data: ContentBlockData }> = ({ data }) => {
+  return (
+    <div
+      className={cn("w-full mb-4 md:mb-0 md:ml-8", {
+        "lg:w-1/4 md:w-1/3": data.aspectRatio === "4by5",
+        "md:w-1/3": data.aspectRatio === "1by1",
+        "lg:w-1/2 md:w-1/3": data.aspectRatio === "16by9",
+      })}
+    >
+      <img
+        src={data.coverImage}
+        alt={data.name}
+        className={cn(
+          "object-cover rounded-lg max-h-none sm:max-h-[500px] mr-auto md:mx-auto",
+          {
+            "aspect-[16/9]": data.aspectRatio === "16by9",
+            "aspect-[4/5]": data.aspectRatio === "4by5",
+            "aspect-[1/1]": data.aspectRatio === "1by1",
+          }
+        )}
+      />
+    </div>
+  );
+};
+
+// Feature Layout
+export const FeatureSection: FC<{ data: ContentBlockData }> = ({ data }) => {
+  const ctas = parseRockKeyValueList(data.callsToAction ?? "");
+
+  return (
+    <section className={cn("content-padding py-16")} aria-label={data.name}>
+      <div
+        className={cn(
+          "flex flex-col md:flex-row gap-10 xl:gap-20 items-center max-w-screen-content mx-auto",
+          {
+            "flex-col-reverse": data.imageLayout === "RIGHT",
+          }
+        )}
+      >
+        {data.coverImage && data.imageLayout === "LEFT" && (
+          <FeatureImage data={data} />
+        )}
+        <div className={`flex-1 flex flex-col gap-4`}>
+          <h2 className="text-text-primary heading-h2 mb-4">{data.name}</h2>
+          {data.subtitle && (
+            <h4
+              className={`text-text-secondary text-lg font-bold uppercase mb-2 tracking-widest`}
+            >
+              {data.subtitle}
+            </h4>
+          )}
+          <HTMLRenderer
+            className="text-text-secondary text-lg"
+            html={data.content}
+          />
+          <div className="flex flex-wrap gap-2 mt-10">
+            {ctas.slice(0, 2).map((cta, idx) => (
+              <Button
+                className="font-normal"
+                intent={idx === 0 ? "white" : "primary"}
+                key={idx}
+                href={cta.url}
+              >
+                {cta.title}
+              </Button>
+            ))}
+          </div>
+        </div>
+        {data.coverImage && data.imageLayout === "RIGHT" && (
+          <FeatureImage data={data} />
+        )}
+      </div>
+    </section>
+  );
+};
